server/requirements.txt
absl-py==0.12.0
bert-extractive-summarizer==0.7.1
blis==0.7.4
catalogue==2.0.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.5
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
flask-ngrok==0.0.25
flatbuffers==1.12
flax==0.3.3
idna==2.10
itsdangerous==1.1.0
jax==0.2.13
jaxlib==0.1.65
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
msgpack==1.0.2
murmurhash==1.0.5
numpy==1.20.2
opt-einsum==3.3.0
packaging==20.9
pathy==0.5.2
Pillow==8.2.0
preshed==3.0.5
pycparser==2.20
pydantic==1.7.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.6.3
sentencepiece==0.1.95
six==1.16.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
thinc==8.0.3
threadpoolctl==2.1.0
tokenizers==0.10.2
tqdm==4.60.0
transformers==4.5.1
typer==0.3.2
urllib3==1.26.4
wasabi==0.8.2
Werkzeug==1.0.1
youtube-transcript-api==0.4.1
torch
